Wave to another py file
