+++ signalworks/tracking/__init__.py
@@ -1,2 +1,3 @@
 # -*- coding: utf-8 -*-
-from .tracking import Event, Track, TimeValue, Wave, Partition  # noqa: F401
+from .tracking import Event, Track, TimeValue, Partition  # noqa: F401
+from .wave import Wave
